refactor(resource-directory-hook): rename
