+++ src/hooks/directoryHooks/useCreateResourceDirectory.ts
